change the E and B
I have removed the huge N^3xN^3 matrix of pointers to calculate the energy and the Bfield.
The command nullify(matrix of pointers) is actually associating the pointers and was taking a LOT of memory.

This has now been completly removed and the huge matrices have now the size (N^3*Nb_neighbours). The index of the neighbors are kept in the %line(:,:) matrices