energy density destribution
the energy density distribution has been coded and checked.
The electric field effect on DMI is coded and checked
The energy density distribution is now decomposed by shell and not by interaction because it is simpler. The user is in charge of changing the interaction by hand to have the energy that he wants.