Add anisotropy magnitude input (fix #61)
Adds 4th real value to magnetic_anisotropy which determines the magnitude and thus the sign of the anisotropy term.
Of the entries 1:3 only the normalizes vector is used for the unit-vector direction.

Alternatively, the old 3-real vector form now uses the norm of the 3-reals as magnitude with the sign of the first non-zero entry.