Fix get_eff_field_single as appearing in #67
In Hamiltonian_derivative/deriv_rank2.f90 the single effective field evaluation routines were mixed up.
The matrix was multiplied to the left instead of the right and reverse.
This resulted in errors including non-symmetric Hamiltonians as the Zeenman-term.
While this fixes the difference between get_eff_field_single and get_eff_field I am not sure this fixes the entire issue.