symmetry determine DMI
I have coded a symmetry finder.
It is based on the code of Jülich and quite primitive. It is mainly used to rotate the generalized exchange tensor to align the DMI along the good axes. If it does not find a symmetry operator, if simply finds an axis and an angle and rotate the tensor - When this happens, the user must check that the DM makes sense.